modify

+++ src/data_cell/sparse_vector_datacell.h
@@ -70,7 +70,7 @@ public:
         size_t io_size = (new_capacity - this->max_capacity_) * max_code_size_ + current_offset_;
         this->max_capacity_ = new_capacity;
         uint8_t end_flag =
-            127;  // the value is mea ingless, only to occupy the position for io allocate
+            127;  // the value is meaingless, only to occupy the position for io allocate
         this->io_->Write(&end_flag, 1, io_size);
         this->offset_io_->Write(&end_flag, 1, new_capacity * sizeof(uint32_t));
     }
@@ -137,7 +137,7 @@ private:
     Allocator* const allocator_{nullptr};
     std::shared_ptr<MemoryBlockIO> offset_io_{nullptr};
     uint32_t current_offset_{0};
-    uint32_t max_code_size_{0};
+    uint64_t max_code_size_{0};
     std::mutex current_offset_mutex_;
 };
